requirements.txt
pytest==6.2.5
PyYAML==5.4.1
subprocess32==3.5.4
libvirt-python==7.6.0
